Update styling for new fuse page
